virtual display interrupt from timer
Bug fix for #348: Booting hvm without monitor will cause TDR

Sometimes display interrupt is needed by hvm although physical
interrupt is disabled that time. It does not happen often but do
happen for below cases:

1, Dom0 boot without any monitor. In such cases, all the pipes are
disabled and ther are no display interrupt generated like VBLANK. When
booting win7 hvm in such case, however, we found that the the driver
tries to enable a pipe and then wait for vblank. If there are no
interrupts coming, the OS wil TDR and BSOD. The code is for VGA
monitor detection.

2, When doing the mode setting, dom0 will disable pipe and then enable
later. Although it is very short time, it still has the potential
issue for hvm to generate TDR for not getting display interrupts.

In order to solve this, virtual display interrupt from timer is
introduced. The virtual interrupt is only injected when hvm has a pipe
enabled but actually the pipe is physically disabled.

Right now we check the physical pipe status by reading the display
owner's vgt pipeconf register, instead of reading from real hardware
for better performance.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>