Add more timeout waits after 1st wait timeout
In some extreme scenario, 500ms timeout is not enough to wait
for ring empty, e.g. when running Heaven with 1920 resolution.
This is one design drawback, that we wait for ring idle in a
kernel thread. Soon we'll haev an improvement to have ring
empty check in interrupt handler, so that main thread no need
to wait for ring empty when receiving ctx switch request.

for now, add a workaround to have at most 3 more timeout waits,
based on whether ACTHD is still moving.

Signed-off-by: Kevin Tian <kevin.tian@intel.com>