Pipe remapping for multi-monitor - part1
HVM and dom0 may have different pipe/port connection.
It is ok when there is only one monitor connected.
but issue will appear when there are two or more monitors connected.
the primary surface will be exchanged because of pipe/port mismatch.

the solution is to create a mapping between the virtual pipe and
the physical pipe.

here is what par1 does:
1,When one transcoder is enabled by the guest, rebuild the pipe mapping.
2,Store the pipe mapping in vgt_device

Signed-off-by: Xiaowei Wang <xiaowei.wang@intel.com>