Remove the useless ASSERTION in auxch_read
The aux_ch_read virtual handler has a very old assertion to ensure
that when in aux_ch_read, the EDID information has been probed and
available. That actually depends on driver's behavior and may not
always be true. So remove it and when it happens, driver just got
some uncertained value.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>