Not allow MI_DISPLAY_FLIP to set inconsistent stride
Bug fixing for #320, which has been verified. Other similar bugs
could be closed also after verify, since they look like the same
root cause: #350, #312, #307

When a hvm is not display owner, if its MI_DISPLAY_FLIP command is
trying to change the physical stride to some value from what display
owner sets, we will make the command to be NOOP. This usually happens
when hvm is running with different resolution.

The reason to do the NOOP is that the stride value is related to the
mode settig on system. If the stride setting is too small, not
consistent with HSYNC, the cmd buffer may hang on wait_for_event for
MI_DISPLAY_FLIP, because the data reading speed cannot match timing.

Only primary plane has such issue. sprite plane should be fine.

Another option other than NOOP is to use system original stride
instead of the value from hvm. But that has a potential issue. If
system stride is larger than hvm stride, there could be out of range
frame buffer surface access, and potentially could happen GTT page
fault etc, although the possibility is very low.

In furutre when panel fitting feature for hvm is checked in, the patch
needs to be changed to check panel fitting setting to decide the NOOP
policy.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>