handle frame counter for virtual vblank
When we are using virtual vblank from timer, the frmcounter MMIO
needs to be virtualized. The read of this MMIO should return different
value although physical MMIO keeps the same.

display owner does not need this since it can access physical display
MMIOs and does not need virtual vblank anytime.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>