Virtualize fdi auto link training in hvm properly
FDI link training is needed for CRT monitor. In HSW, DDI Port E
connects to PCH Display, and auto-training is supported.

This commit handles the auto-training done bit for the FDI auto
link training by referring to mode setting sequence in BSpec. The
bit will be set when the link traing starts, and will be cleared
when DDI port E is disabled.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>