can't figure out eDP input sel when disable pipe
when HVM disable the pipe of eDP. it will clear the bit of eDPInput
Select when it write _REG_TRANS_DDI_FUNC_CTL_EDP.
so we can not figure out the eDP input select from this mmio write.

we can figure out the eDP input select from previous cached vreg.

Signed-off-by: Xiaowei Wang <xiaowei.wang@intel.com>