vgt: idle_render_engine: wait more time when no ACTHD increase in 50ms
The current logic is: when the first 500ms wait times out, we wait 50ms: if
there is no ACTHD increase in the 50ms, we just "break" and "return false"
and finally vgt_do_render_context_switch() will trigger ASSERT(0).

We can fix the logic by not breaking so eagerly.

This patch can work around bug #377, #383, #393 and probably #364(I believe
bug #364 can't appear with this patch).

Signed-off-by: Dexuan Cui <dexuan.cui@intel.com>