Add commnad MI_WAIT_FOR_EVENT handler
MI_WAIT_FOR_EVENT will be used to wait for MI_DISPLAY_FLIP to complete.
so it should be patched to NOOP if the VM is not foreground VM as
MI_DISPLAY_FLIP does.
the pipe this commnad is waiting for should be mapped to the real pipe.

in this patch, the following new define is added:
+#define   MI_WAIT_FOR_PLANE_C_FLIP_PENDING      (1<<15)
+#define   MI_WAIT_FOR_PLANE_B_FLIP_PENDING      (1<<9)
+#define   MI_WAIT_FOR_PLANE_A_FLIP_PENDING      (1<<1)

+#define   MI_WAIT_FOR_SPRITE_C_FLIP_PENDING      (1<<20)
+#define   MI_WAIT_FOR_SPRITE_B_FLIP_PENDING      (1<<10)
+#define   MI_WAIT_FOR_SPRITE_A_FLIP_PENDING      (1<<2)

There is no such define in i915, however we can find them in public IVB spec
from the following location:

IVB - Volume 1 Part 3: Graphics Core - Memory Interface and Commands for the Render Engine

1.2.23 MI_WAIT_FOR_EVENT ...................................................... 193

Signed-off-by: Xiaowei Wang <xiaowei.wang@intel.com>