FRMCOUNT virtualization
Provide hvm the consistent view of FRMCOUNT with the virtual pipe
status. When virtual pipe is enabled, no matter physical pipe is
enabled or disabled, the reading of the counter will reaturn
continuous increasing value according to vblank.

The implementation could be thought this way: FRMCOUNT is a monotonic
increasing counter, which is per-vm and per-pipe. The monotonic
counter increases either by vblank interrupt when hw pipe is enabled,
or by the virtual timer when hw pipe is disabled. A per-vm per-pipe delta
value holds the offset between virtual counter and hardware counter.

Since when virtual pipe is disabled, we do not need calculation actually, the
implementation could be optimized this way: when virtual pipe is first
disabled, we record frmcount value and keep unchanged; when virtual pipe is
first enabled, we calculate the delta value and get the frmcounter as follow:

	frame_counter = hw-counter + delta

We also need to consider the VM's pipe mapping since here we are accessing
hardware registers. The hardware counter should be from the physical pipe that
is mapped to the VM's virtual pipe. When the mapping is not available, the
counter will be from delta only.

S3 resume could be supported also. Since S3 suspend will disable gfx hw and resume will
enable it again, during the enabling in gfx resume, the delta value for frmcount will
be recalculated based on that time hw value and virtual value. Then hvm could still get
the monotonic value.

There is one new macro:

"#define VGT_PIPECONFPIPE(pipeconf) _VGT_GET_PIPE(pipeconf, _REG_PIPEACONF, _REG_PIPEBCONF)"

defined in reg.h. The pipeconf is public available MMIO. The PIPECONF definition could be
found in i915 driver code, and IVB PRM, Vol3 Part3, section 5.1.3.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>