Not really free memory in EDID/DPCD clearing
The previous implementation will free memory in clearing EDID or
DPCD function call. In the case of very frequent display monitor
hotplug, there will be lots of kmalloc and kfree.

This commit introduces a "data_valid" flag for EDID and DPCD data
block. Then when clearing EDID or DPCD during the monitor information
propagation for hotplug out, we could simply change the flag, instead
of real freeing memory. The memory could be reused in next hotplug in
of the same type of monitor.

It could have some side-effect of memory usage in XenGT. Imagine
users first connect DP monitor and then plugout and replace it with
a HDMI monitor. The new code will keep the EDID/DPCD for DP until
the hvm is released. But looks OK since it does not happen often.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>