assign aux channel interrupts to dom0
aux channel read/write is bypassed to hardware for dom0.
as a result, the associated interrupts should be passed to dom0 as well.

this can fix bug 458
[3.11.6-vgt+][Display-Interface][dom0]dp can not boot up in Xmode directly

root cause:
because Aux Channel interrupts are not injected into dom0, i915 driver waits
10 ms for every Aux transaction, it consumes about 10 seconds totally for
DP EDID read. as a result, the display is not ready when lightdm starts up.

Signed-off-by: Xiaowei Wang <xiaowei.wang@intel.com>