Fix deadlock in VGT framebuffer notification
As reported by Dexuan, my simple kernel->userspace notification via sysfs
polling was causing deadlocks under certain loads.

This deadlock was reliably reproduced with the "Prove lock correctness"
debug kernel config option.

This review reworks the notification mechanism to use one of the
existing mechanisms, specifically "Netlink connector"

The issue was that the sysfs_notify api was not safe to call from
interrupt context, since it could sleep, and be rescheduled.

Connector is safe for this:
 "It can be safely called from softirq context"

Note that in handlers.c, I now send an update for all mapped pipes when
we get a display refresh notification through the shared page from the
guest driver (ie, during single buffer mode)

Signed-off-by: Ben Guthro <benjamin.guthro@citrix.com>