inject virtual flip done when vm is running background
when the VM is running in the background, display flip will be ingored
by XenGT, command MI_DISPLAY_FLIP will be replaced as MI_NOOP.
and the pri_surf_mmio_write will be dropped.
as a result, no primary flip done interrupt is triggered by hardware.
while the guest may still waiting for that interrupt.

sulution: inject an virtual flip done interrupt to the guest
if the display flip command is replaced as NOOP or if the display plane surface
address MMIO write is ignored when the VM is running in the background.

Signed-off-by: Xiaowei Wang <xiaowei.wang@intel.com>