increase dom0 default aperture size to 120MB
the origin dom0 default aperture can not meet Ubuntu's needs.
especialy when two or three monitors are connected at the same time.
the default 64 would cause bug 394 and 423.

so we need to increase the dom0 aperture size.

ctx_switch_rb_page and dummy_area are not needed in XenGT now, so we
do not need to reserve them now.
actually vGT only needs 4M+ aperture size, so 8MB aperture should be
ok for vGT, so we can inscrease the dom0 aperture size to 64+56 = 120MB.

Signed-off-by: Xiaowei Wang <xiaowei.wang@intel.com>