Not to check the display mode in pipe mapping
When building the pipe mapping, original code will check TRANS_DDI_FUNC_CTL
MMIO to know which physical pipe has the same port selection with the virtual
pipe. Beyond checking DDI field, the code will check MODE field also. It is OK
when hvm sees the same monitor as physically connected. But when dom0 has some
virtual monitors presented to hvm (through customizing EDID), we may not be
able to expect the same MODE obtained for the port.

It is not necessary to do the MODE check. So it is removed.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>