Return GMBUS NACK when EDID is not available
That is a missing code when I added "data_valid" flag for EDID data
structure. It will cause guest to still get EDID block but with all
zero value.

Normal display feature or hotplug do not seem to be impacted by this,
since all zero block will be considered to be invalid and ignored. But
it is possible that some bugs are related to this.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>