Use virtual pipe_id for notification
In the notification from kernel to user space, the original code
uses the mapped pipe id. That does not seem to be correct (at least
not necessary, depending on how user space handles it)

The returned value of pipe mapping in XenGT is to tell which physical
pipe is related to the virtual pipe in hvm that they connects to the
same port. The information is necessary for direct display model so
that correct PLANE MMIO could be set for a physical port.

In the notification case, actually the user space should try to get
the corresponding FB changes from the message. Then virtual pipe id
should be used to find the related surface base register.

The logic in gem_vgtbuffer could be simplified as well. We do not need
to consider the scenario that the pipe mapping is not yet valid. Only
checking virtual enabling status is good enough.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>