Remove pipe mapping and add port information in fb_decoder
The port information can help users to find the frame buffer for a
given port. The pipe mapping information seems to be usable only
for direct display model. So it is removed.

User can still know the virtual pipe to physical pipe mapping if
wanted, by checking which physical pipe is connected to the port
that corresponds to the virtual port.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>