fix 640x480 issue on some DP monitors
on some DP monitors, the max resolution of the guest would be 640x480.
the root cause of this issue is that XenGT does not probe EDID correctly.

XenGT probes a wrong EDID and presents this wrong EDID to the guest,
the guest OS then failed to get the right edid and then set a lowerst mode.

per DP spec, the 4th byte of aux data specify the length of data to read/write to i2c.
and msg[3] = data length - 1. this means that if we would like to read/write 1 byte
to i2c , we need to program msg[3] = 0 intead of 1.

Signed-off-by: Xiaowei Wang <xiaowei.wang@intel.com>