New sysfs interface for port info/status changes
Co-authored with John Baboval <john.baboval@citrix.com>

New sysfs interface is introduced to change port information, like
the EDID connected to that port, the physical mapped port of the port.
The change will not take effect immediately. Only hotplug or flush
operations, which are all through the writing to "connection" node,
can make the previous setting to work.

The sysfs structure implementation are mostly done by John. Zhiyuan
did more debugging based on that for the hotplug handling and hvm
creation handling.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>
Signed-off-by: John Baboval <john.baboval@citrix.com>