do not detect eDP when resume from s3
THIS IS A TEMP SOLUTION.
XenGT can not detect physically eDP when resume from s3.
the reson why eDP disappears remains unkown so far.
we can workaround this issue by not detecting eDP after s3.
this workaround is safe because eDP does not support hotplug.

this patch can fix bug 550.

the long term solution is to move all EDID probing to user mode.

Signed-off-by: Xiaowei Wang <xiaowei.wang@intel.com>