Move EDID probing into work queue for hot plug
It is a known issue talked earlier this month. Currently we need to hold i915
mutex lock while doing EDID probling. Dexuan found that sometimes the waiting
for i915 lock may hang the system. The reason is that the EDID probing is from
vgt_thread. When it is blocked, the vgt_thread cannot response system
interrupts. Unfortunately there is the case that i915 is waiting for interrupt
while holding the mutex. The solution is to move the probing out of vgt_thread,
instead, a work queue is used.

Concurrency is handled this way. There is a bit set to record the record from
vgt_thread, and the data is protected by a mutex. uevent handler will set it
and work_queue handler will clear it.

bug "#542 [3.11.6-vgt+][win7]get tdr when hotplug dp" should be able to be
fixed by this patch, although it is hard for me to reproduce locally.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>