check edid checksum after probing edid
in some cases, edid from hardware might be wrong or incomplete.
for example, when user plug out one monitor slowly, we might get
parts of edid correctly, but the other parts might be invalid.

we need to add a check to verify that the edid is valid.
if the edid is not valid, we assume the monitor is not connected.

Signed-off-by: Xiaowei Wang <xiaowei.wang@intel.com>