Fix a typo, which will have out-of-range access
It is a bug exposed when Zhi builds XenGT with gcc 4.8. Gcc gives
the warning for the out of range access inside a loop. It is
because a wrong macro used for the loop. The code will only be
run when system is shutting down. So normal tests did not expose
the problem.

Awesome gcc!

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>