vgt: fix some errors in show_batchbuffer().
From Yu:

There're 2 places about show_batchbuffer() I guess are incorrect:

1> In show_ringbuffer(), the show_batchbuffer() is called
incorrectly:
the order of parameter is not consistent with the prototype
of show_batchbuffer().

2> In show_batchbuffer(), the ip_va is translated for gma,
ranging from addr-bytes, to addr+bytes. If the value of addr
is a very smaller one, e.g. 0, the addr-bytes would be somewhat
like 0xffffxxxx(biger than 2G). Hence the vgt_gma_to_va() may be
called to translate some invalid gma, and a fault might be
generated when accessing the ppgtt table.

Special thanks to Yu. This patch can fix DOM0 hang in BUG 524#.

Signed-off-by: Zhi Wang <zhi.a.wang@intel.com>