Fix the EDID/DPCD probing after S3 resume
It is a regression bug. The previous commit introduced mask for
probing after S3 resume. But the parameter for probing is port
number actually. That commit caused only DP_D port to be probed
after s3 resuming.

The issue is fixed in simple way. Considering the probing function
will be removed soon, that should be fine.

This commit can fix bug #402