Fix macro misuse in vgt_decode_sprite_plane_format()
This patch fixes bug 562 - domain 0 will crash when windvd9 play MPEG2
file with hardware decode acceleration. The root cause lies in
vgt_decode_sprite_plane_format(), which will be called when sprite control
register is written. In this routine, the sprite base is calculated with an
incorrect macro - VGT_SPRSURFPIPE. Replacing it with VGT_SPRSURF will fix it.

Signed-off-by: Yu Zhang <yu.c.zhang@intel.com>