Fix the gpu hang issue described in bug 543
There's a weird gpu hang issue described in bug 543. When windows
hvm is recreated repeatedly, the gpu may hang at the first command,
unless RC6 is disabled.
This commit is a workaround to fix this: __vgt_rendering_restore()
is modified to do the post-read each time a mode control register
is restored.

Signed-off-by: Yu Zhang <yu.c.zhang@intel.com>