calculate drm name dynamically
drm name <-> port name is not fixed actually.
the mapping depends on whether the physical port is present or not
we should figure out whether PORT_B, C, D are present or not,
and then cacualte the mapping.

add one sysfs "presence" to indicate whether the port is present or not
under each physcal PORT.

refer to intel_setup_outputs() in i915 driver.

Signed-off-by: Xiaowei Wang <xiaowei.wang@intel.com>