DAC_CTL(ADPA_CTRL) virtualization policy changes
This commit changed the virtualization policy for DAC_CTL/ADPA_CTRL(0xE1100).
The original code will fully virtualize the "force trigger" and the "channel
status bits" (bit 24-25) according to the monitor connection information.

It is doable when vgt does the EDID probing since vgt knows exactly such
monitor connection information. However, with the probing being moved to user
space, such virtualization is not feasible.

Another change in this commit is: not to write the MMIO in physical CRT hotplug
handler. The original code treats the "status bits" as sticky bits, then clear
it in physical handler. But actually those bits are not sticky bits. Such write
is not necessary.

Summary of the changes:

1, always return physical value of DAC_CTL for display owner
2, not to write DAC_CTL in physical crt hotplug handler

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>