Fix a typo, which caused EDID not read out from HDMI monitors
The port information was not saved from the gmbus0 write, which will
cause the later gmbus operations failed to get the EDID info. The previous code
was only tested with VGA and DP, having hoped to cover GMBUS with VGA, but
obviously not.

This commit can fix #628: "cannot display xengt vm in SDP system"

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>