Send display detection uevent in hotplug virq
The old logic did this in i915 mode setting. That has problems in
3.17 kernel. When physical hotplug happens, i915 mode setting may
not always be triggered.

The new code sends monitor detection in hotplug interrupts, which
is more meaning.

Meanwhile, there is no correctness issue. The previous concern is
that detection uevent cannot be too earlier than monitor information
ready in dom0. But that is not a problem. When user space calls i915
sysfs interface to check the "state" of a PORT, the port detection
will be triggered that time. So the information obtained is always
up-to-date.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>