vgt: disable i915 render state under VGT environment.
Render state in i915 is broken, it dispatches a PPGTT
batch buffer before PPGTT is enabled, I tried to fix it
but then another issue is found: MI_BATCH_BUFFER_END is
missed in i915 render state array. It should be very very
lucky that it works every time before.

Signed-off-by: Zhi Wang <zhi.a.wang@intel.com>