vgt: introduce irq map.
irq map is a table which used to describe the relationships
between irq groups. For example, on HSW, when bit 28 of DE_IIR
is set, it indicates PCH_IIR has pending events. And a irq map
describes this relationship like:

{ PCH_IRQ, IRQ_INFO_PCH, ~0 },

Upstream IRQ event bit PCH_IRQ is associated to the whole
registers bits of downstream IRQ group IRQ_INFO_PCH.

On BDW, there is a special case: an upstream IRQ event bit
can be associated to only a part of registers bits of
downstream IRQ group, e.g. GT0 and GT1, which can be
managed by the bitmask field here.

Signed-off-by: Zhi Wang <zhi.a.wang@intel.com>