vgt: split/abstraction between vgt core and hypervisor logic - stage I
As a preparation step to move vgt code to i915 and better support
of xen/kvm simultaneously. All Xen specific interfaces are in
hypercall.c which supposes to be a vgt.c in driver/xen/ in the
future. Now no registration mechanism is provided, just a hardcoded
callback to xen interfaces. all hypervisor specific interfaces
are wrapped into hypervisor_xxx to be used by common vgt code.

It's only the stage I of this effort - an abstraction skeleton.
There are still Xen-specific operations such as the irq injection
method, are not abstraced in this stage.

The final goal should be remove all xen header inclusions and
references in common vgt code so we can move common code to
i915.

Signed-off-by: Kevin Tian <kevin.tian@intel.com>
Signed-off-by: Jike Song <jike.song@intel.com>