drm/i915: i915 host irq mediation
With this patch, the hardware interrupt resources are under management
of vgt, the entrance of service routine is vgt_interrupt().

Whenever vgt decides to deliver an virtual interrupt to host i915
driver, it calls vgt_host_irq(). An irq_work is pushed into the end
of its queue, and delivered to host i915 via IPI - an encapsulation
provided by linux kernel, which is very similar to our mechanism
used previously.

To clarify:

	1) the hardware interrupt is processed by vgt_interrupt
	2) the i915 interrupt handler is called in hardirq context
	3) currently the irq_work is schedule on the same cpu where
	   the enqueue action took place.

Signed-off-by: Jike Song <jike.song@intel.com>

Conflicts:
	drivers/gpu/drm/i915/i915_drv.c
	drivers/gpu/drm/i915/i915_irq.c