Add new write-protection interfaces
A new write-protection interface, hvm_wp_page_to_ioreq_server(),
is defined to set or clear the write-protection status for one
page. To write-protect one memory range, this range is first added
into the rangeset of the ioreq-server; then a special memory type,
HVMMEM_mmio_write_dm, is set for this range.

Signed-off-by: Yu Zhang <yu.c.zhang@intel.com>