vgt: display: redesign the drm name getting in vgt_mgr
current vgt_mgr implementation assume that the port connection will follow
the order that the front port like PORT_B have higher priority to enable
any type of connection. However it's not true, e.g. it cannot handle the case
card0-DP-1 connect to PORT_C. Actually the connector type name and port mapping
depend on the setting in VBT.

This change export the port and connector type mapping in i915 debugfs
and redesign the drm name getting policy in vgt_mgr base on the mapping.

Signed-off-by: Ping Gao <ping.a.gao@intel.com>