Leverage ioreq-server to do io forwarding.
Ioreq-server is used to forward the pio/mmio accesses from HVM
to vgt device model. Several new hypervisor interfaces are added
to create/enable/destroy an ioreq server.
Old interfaces such as vgt_hvm_enable() are removed.
Besides, the vgt_hvm_write_cf8_cfc() and vgt_hvm_read_cf8_cfc()
are updated because we do not need to keep track of the last cf8,
which is recorded in the hypervisor side now.

V2:
1> rename vgt_hvm_read/write_cf8_cfc to vgt_hvm_read/write_cfg_space;
2> rename hvm_enable_iorequest_server to hvm_toggle_iorequest_server;
3> add error handling in map_hvm_iopage().

Signed-off-by: Yu Zhang <yu.c.zhang@intel.com>