i915/vgt: workaround for bug#685 Dom0 S3 becomes slowly sometime
In current vgt driver kernel thread vgt_thread is set to freezable, so
in S3 process it will be frozen before i915 driver suspend callback being
invoked, which cause the page flip interrupt fail to deliver to i915
driver and block the Dom0 S3. This is a workaround to double check the
page flip status instead of only waiting for page flip interrupt.

This issue will disappear after the vgt driver refactor and integrate
with i915 driver.

Signed-off-by: Yulei Zhang <yulei.zhang@intel.com>